frontends/octoshop_docker/requirements.txt
streamlit==1.24.1
Pillow==9.5.0
numpy==1.24.4
octoai-sdk==0.7.0
